+++ app/(navigatable)/(projects)/projects/create/primary-data-source-selector.tsx
@@ -1,11 +1,6 @@
 'use client'
 
 import React, { useEffect, useState } from 'react'
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from '@radix-ui/react-collapsible'
 import { getOrgStatus } from 'app/react-utils'
 import { SaraSession } from 'auth'
 import { Button } from 'components/ui/button'
@@ -16,21 +11,22 @@ import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  DropdownMenuLabel,
   DropdownMenuTrigger,
 } from '../../../../../components/ui/dropdown-menu'
 import { getResource } from './../../../../../app/saraClient'
 import LoadingSpinner from './../../../../../components/loading-spinner'
 import { Org, type GitHubRepo } from './../../../../../lib/data-model-types'
 import { InfoCircledIcon, LockClosedIcon, LockOpen2Icon } from '@radix-ui/react-icons'
-import { Callout } from '@radix-ui/themes'
+import { Badge, Callout } from '@radix-ui/themes'
 import Link from 'next/link'
 
 interface DataSourceSelectorProps {
+  userIsPremium: boolean
   setControlledProjectDataSources: (gitHubRepos: GitHubRepo[]) => void
 }
 
 const PrimaryDataSourceSelector = ({
+  userIsPremium,
   setControlledProjectDataSources,
 }: DataSourceSelectorProps) => {
   const session = useSession()
@@ -142,7 +138,7 @@ const PrimaryDataSourceSelector = ({
               )}
           </Button>
         </DropdownMenuTrigger>
-        <DropdownMenuContent sideOffset={8} align="start" className="min-w-64 max-h-60">
+        <DropdownMenuContent align="start" className="min-w-64 max-h-60">
           {orgs.map((org: Org) => (
                 <DropdownMenuItem
                   key={org.name}
@@ -175,7 +171,7 @@ const PrimaryDataSourceSelector = ({
                     )}
                 </Button>
               </DropdownMenuTrigger>
-              <DropdownMenuContent sideOffset={8} align="start" className="min-w-64 max-h-60">
+              <DropdownMenuContent align="start" className="min-w-64 max-h-60">
                 {githubReposForOrgs[selectedGithubOrg?.name] ? 
                     githubReposForOrgs[selectedGithubOrg.name].map((repo: GitHubRepo) => (
                         <DropdownMenuItem
@@ -184,15 +180,19 @@ const PrimaryDataSourceSelector = ({
                             setSelectedGithubRepo(repo)
                             setControlledProjectDataSources([repo])
                             }}
+                            disabled={repo.private && !userIsPremium}
                         >
-                            <span className="ml-2 text-ellipsis whitespace-nowrap overflow-hidden">
-                            {repo.name}
+                            <span className="mx-2 text-ellipsis whitespace-nowrap overflow-hidden">
+                              {repo.name}
                             </span>
-                            {repo.private ? 
-                              <div title="Repo Unlocked" className="ml-1">
-                                <LockOpen2Icon className="w-4 h-4" />
-                              </div> 
-                            : null}
+                            {(repo.private && !userIsPremium) && 
+                            <div className="p-1 rounded text-red-500 font-semibold text-xs bg-red-100">
+                              Premium Requred
+                            </div>}
+                            {(repo.private && userIsPremium) && 
+                            <div title="Premium Access" className="px-1 rounded text-green-500 font-semibold text-xs bg-green-100">
+                              P
+                            </div>}
                         </DropdownMenuItem>
                         ))
                         : <DropdownMenuItem>